// src/screens/auth/signup/registerNewCompany/types.ts
export type IRegisterNewCompanyProps = {
  companyName: string;
  companyNameError: string;
  companyEmail: string;
  companyEmailError: string;
  contactNumber: string;
  contactNumberError: string;
  address: string;
  addressError: string;
  location: string;
  locationError: string;
  industry: string;
  industryError: string;
};